#41 | accounts dir structure refactoring finished
