+++ src/hooks/accounts/resolvers/query/activeAccount.tsx
@@ -1,10 +1,10 @@
 import { useCallback } from 'react';
 import { Account } from '../../../../generated/graphql';
-import { useResolverToRef } from '../mutation/useAccountsMutationResolvers';
 import { GET_ACCOUNTS } from '../../queries/useGetAccountsQuery';
 import { usePersistActiveAccount } from '../../lib/usePersistActiveAccount';
 import { find } from 'lodash';
 import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
+import { useResolverToRef } from '../useAccountsResolvers';
 
 // make sure the __typename is well typed
 export const __typename: Account['__typename'] = 'Account';
