+++ src/hooks/assets/resolvers/useGetAssetsQueryResolver.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { Asset } from '../../../generated/graphql';
-import { useResolverToRef } from '../../accounts/resolvers/mutation/useAccountsMutationResolvers';
 import { useGetAssets } from '../useGetAssets';
+import { useResolverToRef } from '../../accounts/resolvers/useAccountsResolvers';
 
 export const __typename: Asset['__typename'] = 'Asset';
