+++ src/hooks/config/useConfigQueryResolvers.tsx
@@ -1,12 +1,12 @@
 import { ApolloCache, NormalizedCacheObject } from '@apollo/client';
 import { useCallback } from 'react';
-import { useResolverToRef } from '../accounts/resolvers/mutation/useAccountsMutationResolvers';
 import {
   GetActiveAccountQueryResponse,
   GET_ACTIVE_ACCOUNT,
 } from '../accounts/queries/useGetActiveAccountQuery';
 import { usePolkadotJsContext } from '../polkadotJs/usePolkadotJs';
 import { usePersistentConfig } from './usePersistentConfig';
+import { useResolverToRef } from '../accounts/resolvers/useAccountsResolvers';
 
 export const accountCurrencyMapDataType = 'Option<u32>';
