+++ src/hooks/vesting/useVestingMutationResolvers.tsx
@@ -1,5 +1,4 @@
 import { useCallback } from 'react';
-import { useResolverToRef } from '../accounts/resolvers/mutation/useAccountsMutationResolvers';
 import { usePolkadotJsContext } from '../polkadotJs/usePolkadotJs';
 import { web3FromAddress } from '@polkadot/extension-dapp';
 import { ClaimVestedAmountMutationVariables } from './useClaimVestedAmountMutation';
@@ -12,6 +11,7 @@ import {
   GET_ACTIVE_ACCOUNT,
 } from '../accounts/queries/useGetActiveAccountQuery';
 import { ApiPromise } from '@polkadot/api';
+import { useResolverToRef } from '../accounts/resolvers/useAccountsResolvers';
 
 /**
  * Run an async function and handle the thrown errors
