+++ src/hooks/vesting/useVestingScheduleQueryResolvers.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { Account } from '../../generated/graphql';
-import { useResolverToRef } from '../accounts/resolvers/mutation/useAccountsMutationResolvers';
 import { useGetVestingScheduleByAddress } from './useGetVestingScheduleByAddress';
+import { useResolverToRef } from '../accounts/resolvers/useAccountsResolvers';
 
 export const useVestingScheduleQueryResolvers = () => {
   const getVestingScheduleByAddress = useGetVestingScheduleByAddress();
